Removed the usage of Serial with this version of the Arduino code, since I'm not testing that code any longer. Migrated some code from the main project into a new Library project; I'll be putting all of the light manipulation and device communication in there.
Added a Moonlight control and have begun to replace the old code with the new concepts that will be setting LEDs across all the strips. Next step will be to migrate to an MVVM pattern so that the ActiveManager can be swapped in and out, and will likely begin to use bindings at that point rather than a Start button approach.

At high speeds, the application was killing CPU usage, so I've switched to an extension method that does fast writing instead, mostly based on a comment in a issues thread on HidLibrary.