Revising the communication between the Teensy and the Application to use control codes, even for setting individual pixels, so as to have a consistent process. Adding a universal setting to greatly the reduce the packets that have to be sent for things like moonlighting or single colors.
Due to the communication changes, removed the PackageGenerator and replaced it with a couple of calls in LEDs.

Tweaked how the HidExtensions work based on how the real version works, forcing it to copy the buffer into a second array that's based on the size of the array used by the device. FastRead probably isn't fully baked, but I'm likely to remove it at some point anyways; I only need high throughput from the computer to the Teensy, not the other way around.

Added the MVVM Light library, to start migrating to using MVVM instead.