Added settings tracking so that the application remembers the last settings chosen for most of the light managers. Moved the light managers into their own folders for both the data templates and the ViewModels folder.
Added new OffViewModel and ManualViewModel options. Updated Gradient to use a new Manager; now only Off and SingleColor interact with the package generation and sending directly; everything else goes through a model.

Fixed a bug in the SendRange design where it didn't start with a fresh byte array on shipping out the previous one, which caused existing messages to remain in the array.

Redesigned everything to use SimpleIOC for dependencies - LEDs is now gone, Communicator is no longer a singleton, etc.).

Otherwise, moved around classes and adjusted namespaces.

On the Arduino side, moved files around, and created the ability to reduce the verbosity of logs from the application. The application-side of that remains to be written, though it does have the constant defined for a Settings message.