Revised Settings concept to be a StatusViewModel instead.  Created the new Status objects for tracking the current state of the device.
Updated Communicator to passive update status, and provide a hook to force a status refresh. Created the connection to send updates to Settings (currently, only logging level) and retrieve the resulting values. Fixed a bug in the FastRead code - it was relying on the user to give it a proper buffer size. It now returns a buffer back to the requester rather than asking them to new it up.

On the Arduino side, it finally dawned that if I create my DEFINEs in the main file, then they're available for all of the other files. Also added the ability to update the settings and return the status, tweaked the definition for the SendMessage stuff for talking to the Serial bus, and moved the Processor to a switch(case) instead of if/else, because, really.